pyproject.toml
[tool.poetry]
name = "docker-swarm-mcp"
version = "0.4.0"
description = "HTTP-based Model Context Protocol server for Docker operations with context-preserving tool gating"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "0.104.1"
uvicorn = {extras = ["standard"], version = "0.24.0"}
docker = "^7.1.0"
pydantic = "2.5.0"
pyyaml = "6.0.1"
urllib3 = ">=2.5.0,<3.0"
requests = ">=2.32.4,<3.0"
jsonschema = "^4.20.0"
pyjwt = "^2.8.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4"
pytest-cov = "^4.1"
mypy = "^1.7"
ruff = "^0.1"
bandit = "^1.7"
httpx = ">=0.24,<0.28"  # FastAPI 0.104.x + Starlette 0.27.x expect httpx<0.28

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
no_implicit_optional = true

[[tool.mypy.overrides]]
module = "docker.*"
ignore_missing_imports = true

[tool.bandit]
exclude_dirs = ["tests", ".venv", "docs"]
skips = ["B101"]
targets = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "--cov=app --cov-report=term-missing --cov-report=html"

[tool.poetry.scripts]
test = "poetry_scripts:test"
test-fast = "poetry_scripts:test_fast"
test-cov = "poetry_scripts:test_cov"
